updated fixture implementation
