+++ .github/workflows/pull-request.yml
@@ -9,7 +9,7 @@ jobs:
   test:
     strategy:
       matrix:
-        version: [ 18, 20 ]
+        version: [ 18, 20, 22 ]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
