+++ CHANGELOG.md
@@ -10,5 +10,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 :pencil: - chore
 :microscope: - experimental
 
+## [0.2.0]
+- :rocket: updated fixture implementation
+
 ## [0.1.0]
 - :rocket: initial implementation
