+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@qavajs/playwright-runner-adapter",
-  "version": "0.1.1",
+  "version": "0.2.0",
   "description": "adapter for playwright test runner",
   "scripts": {
     "build": "tsc",
@@ -11,13 +11,12 @@
   "license": "MIT",
   "dependencies": {
     "@cucumber/gherkin": "^28.0.0",
-    "@cucumber/messages": "^24.1.0",
-    "glob": "^10.3.12"
+    "glob": "^10.3.15"
   },
   "devDependencies": {
-    "@cucumber/cucumber": "^10.3.2",
-    "@playwright/test": "^1.42.1",
-    "@types/node": "^20.12.2",
-    "typescript": "^5.4.3"
+    "@cucumber/cucumber": "^10.7.0",
+    "@playwright/test": "^1.44.0",
+    "@types/node": "^20.12.12",
+    "typescript": "^5.4.5"
   }
 }
